feat #18 : 커스텀에러 코드 적용, 자잘한 버그 수정
