+++ src/main/java/com/backend/soullive_a/controller/ModelController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/api/model")
 public class ModelController {
     private final ModelService modelService;
 
@@ -18,7 +19,7 @@ public class ModelController {
      * @param modelId
      * @return
      */
-    @GetMapping("/api/model/{modelId}")
+    @GetMapping("/{modelId}")
     public BaseResponse<ModelResponse> getModel(@PathVariable Long modelId) {
         return BaseResponse.<ModelResponse>builder()
                 .isSuccess(true)
@@ -34,7 +35,7 @@ public class ModelController {
      * @param request
      * @return
      */
-    @PostMapping("/api/model")
+    @PostMapping("/")
     public BaseResponse<ModelResponse> createModel(@RequestBody @Valid ModelRequest request) {
         return BaseResponse.<ModelResponse>builder()
                 .isSuccess(true)
