+++ src/main/java/com/backend/soullive_a/dto/response/ModelResponse.java
@@ -1,6 +1,5 @@
 package com.backend.soullive_a.dto.response;
 
-import com.backend.soullive_a.entity.ModelIntroduction;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
