+++ src/main/java/com/backend/soullive_a/entity/model/introduction/ModelImageKeyword.java
@@ -1,5 +1,6 @@
-package com.backend.soullive_a.entity;
+package com.backend.soullive_a.entity.model.introduction;
 
+import com.backend.soullive_a.entity.model.Model;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -8,13 +9,17 @@ import lombok.*;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ModelRecentAdvertisement {
+public class ModelImageKeyword {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "MODEL_RECENT_ADVERTISEMENT_ID", nullable = false)
+    @Column(name = "MODEL_IMAGE_KEYWORD_ID", nullable = false)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "MODEL_ID", nullable = false)
     private Model model;
+
+    @Column(name = "KEYWORD", nullable = false)
+    private String keyword;
 }
