+++ src/main/java/com/backend/soullive_a/repository/ModelRepository.java
@@ -1,6 +1,6 @@
 package com.backend.soullive_a.repository;
 
-import com.backend.soullive_a.entity.Model;
+import com.backend.soullive_a.entity.model.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
